+++ base_aux/monitors/m1_imap.py
@@ -6,8 +6,7 @@ import imaplib
 import email
 
 from base_aux.lambdas.m1_lambdas import Lambda
-from base_aux.privates.m0_base import *
-from base_aux.privates.m5_auto import *
+from base_aux.privates.m1_privates import *
 from base_aux.aux_attr.m1_attr2_nest2_lambdas_resolve import NestInit_AttrsLambdaResolve
 from base_aux.alerts.m2_select import *
 from base_aux.alerts.m0_base import *