+++ HISTORY.md
@@ -25,7 +25,7 @@
 0.2.11 (2025/02/24 18:32:33)
 ------------------------------
 - [PRJBase] move into share +use Version  
-- [AttrsAux] add AttrTemplate_Direct/Callable  
+- [AttrsAux] add Nest_AttrKit/Callable  
 
 0.2.10 (2025/02/19 16:33:46)
 ------------------------------